customizing::split app with provider &component
